Prepare test for additional cases.
Adds field names to struct literals. This makes it much easier to add new fields in the future.
'go vet' has a warning for this which suggests always use keyed struct
literals.

Also add some new lines to the testcases to make them easier to read on narrower windows by
making the lines shorter.

Give the testCase type a name, and move the test run function above the test cases. This is
a pattern we have used a bunch in Consul. It can make large table tests easier to read because
the test logic is closer to the start of the function. This allows the
test to be read from top-to-bottom, instead of having to scroll past all
the test cases each time.