Use a Logger interface instead of the stdlib logger
There is a Go idiom that goes

> Accept interfaces, return structs

By accepting a Logger interface the caller can plug in any log system
they want. This is ideal for a library because it allows the library to
be used with whatever logging system the application has already chosen.

Most hashicorp applications use hashicorp/go-hclog, so the interface was
chosen to match https://pkg.go.dev/github.com/hashicorp/go-hclog#Logger
but a caller can just as easily implement this interface with a small
shim around the stdlib logger.

Another reason for choosing the hclog interface is because it supports
structured logging. Structured logs can be printed unstructured,
but unstructured logs can not be structured, so by choosing structured
we can support both.