Add context.Context argument
Context allows us to set a deadline or cancel the operation. This is
particularly important for executables to prevent  them from running too
long.

I can add a test for this in a follow up commit once my other PR (which
updates the test setup) is merged.

This change also has a secondary advantage. On switching from from
net.LookupIP to net.DefaultResolve.LookupIPAddr we no longer have to
convert the IP addreses to IPAddrs.

As we can see from the implemenation of LookupIP
https://cs.opensource.google/go/go/+/refs/tags/go1.17.5:src/net/lookup.go;l=192

LookupIP calls the same underlying function. It also creates a slice of
the IPs. So by making this change we remove two sets of allocations
which converted from []IPAddr -> []IP -> []IPAddr.